Import form the snapshot ecartis-1.0.0-snap20060813
STORY: I tried :mailman, :majordomo and did take a look at some
*non-free* MLM. Mailman is the most update software, but it is too
complex. On Debian, the native package requires me to install Apache
or an HTTP server. Some mailman commands go to the hell without saying
Goodbye to me (I mean I could not find its logs).

For some reason, :majordomo requires a :suid flag. God, I do not want
this for a stable sofware released 13 years ago; who knows what
would happen to that :suid stuff (NOTE: vger.kernel.org uses it.)

So here goes. The old domain ecartis.org is obsolete. The new one is
ecartis.net. The page is said to be updated on 2013-July-07, but on
the same page the :webmaster says the last update is on 2013-Jun-07.
And the source would be released soon on Github. I sent an email to
jtrigg (at) ecartis (dot) net if he would have a CVS tree for me.
No response. Maybe he is too busy. So I went to Ubuntu/Debian dev. page
to download the latest source from them. Launchpad is good they keep
a mirror of the source code. I also checked out ftp.ecartis.org to
see if there is any difference. Okay, source seems not to be *hacked*.

This morning I've found the author released a new version v1.0.1 on his
FTP site ftp://ftp.ecartis.net/pub/ecartis/releases/1.0.1/. This is on
2010. Cool. There is a minor code change since v1.0.0. I wonder if
someone else finds this *egg* :)

There are some things in the code that need to be fixed. For example,
the use of (sprintf) or (strcpy). You can see in TODO page. I think
I would give a try and help. That's way I create this official mirror.