get_number accepts the 2nd parameter (default value)
The primary purpose is to get some default value if there is
something wrong with (get_number). In some cases, the default
value 0 (as in the old behavior) may not be useful. For example,
when get SMTP port number, the default value would be 25.

We also provide some random default values for variables.

Please note that the (log_printf) now uses leve 3 as default.
This would help in the early steps of the program when the
configuration file is not ready, and log level is not valid.
using level 3 will hide a lot of messages, but not all of thems
which are useful for debug (file error, configuration error,...)